Moved strip whitespace method so that it is not duplicated
