+++ app/main/services/conversions.py
@@ -1,0 +1,5 @@
+import re
+
+
+def strip_and_lowercase(value):
+    return re.sub(r'[\s+|\W+]', '', value).lower()
