+++ app/main/services/search_service.py
@@ -1,5 +1,4 @@
 import os
-import re
 
 from elasticsearch import Elasticsearch, TransportError
 from app.mapping import SERVICES_MAPPING
