--- scripts/test-indexing.sh
@@ -1,2 +1,0 @@
-curl -H"Content-type: application/json" -X POST http://localhost:5001/g-cloud/services/123456789 -d'{"service" : {"id":"123456789","lot":"IaaS","serviceName":"This is my service","serviceSummary":"My Service does all these things","supplierName":"John Smith enterprises","serviceBenefits":["benefit 1","benefit 2"],"serviceFeatures":["feature 1","feature 2"],"serviceTypes":["type1", "type2", "type3"]}}'
-curl -H"Content-type: application/json" -X POST http://localhost:5001/g-cloud/services/123456780 -d'{"service" : {"id":"123456789","serviceName":"This is my service","serviceSummary":"My Service does all these things","supplierName":"John Smith enterprises","serviceBenefits":["benefit 1","benefit 2"],"serviceFeatures":["feature 1","feature 2"],"serviceTypes":["type1", "type2", "type3"]}}'
